test(worktree): move inline tests to sibling module
Agend-Agent: archfix-codex-10
Agend-Task: t-20260712133728150265-40783-91
Agend-Branch: fix/release-unpreservable-nested-wip
Agend-Issued-At: 2026-07-12T13:37:54.784195+00:00
